up case

+++ case_study_module4/furuma_management/src/main/java/com/codegym/controller/CustomerController.java
@@ -2,6 +2,7 @@ package com.codegym.controller;
 
 import com.codegym.dto.CustomerDto;
 import com.codegym.model.Customer;
+import com.codegym.model.CustomerType;
 import com.codegym.service.ICustomerService;
 import com.codegym.service.ICustomerTypeService;
 import org.springframework.beans.BeanUtils;
@@ -15,6 +16,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.util.List;
 import java.util.Optional;
 
 @Controller
@@ -25,12 +27,27 @@ public class CustomerController {
     @Autowired
     private ICustomerTypeService iCustomerTypeService;
 
+
+    @ModelAttribute("customerTypeList")
+    public List<CustomerType> getCustomerType(){
+        return iCustomerTypeService.findAll();
+    }
+
+
     @GetMapping("/list")
     public String getListCustomer(Model model, @PageableDefault(value = 3) Pageable pageable,
-                          @RequestParam Optional<String> customerName, RedirectAttributes redirectAttributes) {
-        String keywordVal = customerName.orElse("");
-        model.addAttribute("customerList", iCustomerService.findAllByCustomerNameContaining(keywordVal, pageable));
-        model.addAttribute("keyword", keywordVal);
+                                  @RequestParam Optional<String> customerName,
+                                  @RequestParam Optional<String> customerAddress,
+                                  @RequestParam Optional<String> customerType,
+                                  RedirectAttributes redirectAttributes) {
+        String keywordVal1 = customerName.orElse("");
+        String keywordVal2 = customerAddress.orElse("");
+        String keywordVal3 = customerType.orElse("");
+        model.addAttribute("keyword1", keywordVal1);
+        model.addAttribute("keyword2", keywordVal2);
+        model.addAttribute("keyword3", keywordVal3);
+        model.addAttribute("customerList", iCustomerService.findAllByCustomer
+                (keywordVal1, keywordVal2, keywordVal3, pageable));
         redirectAttributes.addFlashAttribute("message", "");
         return "/customer/list";
     }
@@ -39,7 +56,6 @@ public class CustomerController {
     @GetMapping(value = "/create")
     public String createForm(Model model) {
         model.addAttribute("customerDto", new CustomerDto());
-        model.addAttribute("customerType", iCustomerTypeService.findAll());
         return "/customer/create";
     }
 
@@ -61,6 +77,9 @@ public class CustomerController {
     @PostMapping(value = "/delete")
     public String deleteCustomer(Integer id, RedirectAttributes redirectAttributes) {
         Customer customer = iCustomerService.findById(id);
+        if (customer == null) {
+            return "error";
+        }
         iCustomerService.delete(customer);
         redirectAttributes.addFlashAttribute("message", "delete successfully");
         return "redirect:/customer/list";
@@ -82,7 +101,6 @@ public class CustomerController {
             CustomerDto customerDto = new CustomerDto();
             BeanUtils.copyProperties(customer, customerDto);
             model.addAttribute("customerDto", customerDto);
-            model.addAttribute("customerType", iCustomerTypeService.findAll());
             return "/customer/edit";
         }
